refactor(client): extract BlipTable columns and parts barrel to clear import/max-dependencies
