+++ packages/client/src/components/radar/blipTableParts.ts
@@ -1,0 +1,16 @@
+// The building blocks BlipTable composes: the toolbar, the row/bulk
+// sub-components, the filter/sort helpers, the edit-draft type, and the column
+// defs. Grouping the re-exports here lets BlipTable import them as a single
+// dependency. This barrel only re-exports leaf modules — none of them import it
+// back — so there is no circular dependency.
+export type { BlipEditDraft } from "./BlipEditRow";
+export { buildBlipColumns } from "./blipTableColumns";
+export {
+  ALL,
+  countByField,
+  filterAndSortBlips,
+  NO_CHANGE,
+} from "./blipTableFilters";
+export type { BulkPatch, SortKey } from "./blipTableFilters";
+export { BlipBulkBar, BlipDisplayRow, BlipEditRow } from "./blipTableRows";
+export { BlipTableToolbar } from "./BlipTableToolbar";
